Left-right targets
It used to be that we were simply getting the center of mass through
moments and the X value of that was used as a center goal. This made a
lot of sense when the output was noisy, but nowadays gymnasts are
detected very cleanly. This means that we can get the average between
the leftmost and the rightmost pixels and use that as the goal. The
result is slightly different (better in some cases and worse in
others), but generally it is more robust because it can handle more
than one gymnast on the floor.