Add support for writing to loopback devices
This may be required if there is a need to use the video stream
elsewhere. However, it is better to use gstreamer to duplicate the
stream into two loopbacks and use one of them for image
processing. Doing the same in python may result in poor framerate in
the outgoing stream.